removed excess files, added folders for cleaner file tree, updated font and css properties
