+++ client/src/components/Hero/Hero.css
@@ -1,0 +1,8 @@
+.hero-section-one {
+   background: none;
+   max-height: 350px;
+    }
+
+.hero-section-two {
+    background-color: blueviolet;
+}
