+++ client/src/fonts/fonts.css
@@ -1,4 +1,5 @@
 @font-face {
-    font-family: 'Amarante-Regular';
-    src: url('../src/fonts/Amarante-Regular.ttf') format('truetype');
+    font-family: 'Playfair Display';
+    font-style: italic;
+    src: url('../src/fonts/PlayfairDisplay.ttf') format('truetype');
   }
