+++ client/src/pages/Home/Home.css
@@ -3,3 +3,4 @@
     color: white;
     text-shadow: 2px 2px 2px rgb(170, 142, 80);
 }
+
