+++ client/src/pages/Home/Home.jsx
@@ -1,15 +1,25 @@
 import React from 'react';
-import BoxSystemProps from '../../components/CardDrawing/Card';
 import './Home.css';
+import { Grid } from '@mui/material';
+import HeroSectionOne from '../../components/Hero/HeroOne';
+// import HeroSectionTwo from '../../components/Hero/HeroTwo';
+
 
 const Home = () => {
     return (
-        <div className='flex flex-col justify-center items-center h-screen text-center'>
-                <h1 className='home-font text-3x1 font-bold text-center'>Embark on a Journey of Reflection!</h1>
-                <BoxSystemProps />
-                <h2>Explore various spreads, delve into the meanings of the cards, <br />and gain personal insight.</h2>
-                <h1> Discover what the cards hold for you.</h1>  
-            </div>
+        <Grid
+            container
+            spacing={4}
+            direction='row'
+            justifyContent='center'
+            alignItems='center'>
+            <Grid item>
+                <HeroSectionOne />
+            </Grid>
+            {/* <Grid item>
+                <HeroSectionTwo />
+            </Grid> */}
+        </Grid>
     );
 };
