--- client/src/pages/Login.jsx
@@ -1,18 +1,0 @@
-// import React from 'react';
-// import LoginModal from '../components/Authentication/LoginModal';
-
-// const Login = () => {
-//     return (
-//         <div className='flex items-center min-h-screen '>
-//             <div className='containe mx-auto'>
-//                 <div className='max-w-md mx-auto my-10'>
-//                     <div className="m-7">
-//                         <LoginForm />
-//                     </div>
-//                 </div>
-//             </div>
-//         </div>
-//     );
-// };
-
-// export default Login;
